+++ src/main/ts/ephox/alloy/keying/AcyclicType.ts
@@ -1,10 +1,7 @@
-import { SugarEvent } from '../alien/TypeDefinitions';
-import { EventFormat, SimulatedEvent } from '../events/SimulatedEvent';
-import { AlloyEventHandler } from '../api/events/AlloyEvents';
-import { AlloyComponent } from '../api/component/ComponentApi';
+import { FieldSchema } from '@ephox/boulder';
+import { Fun } from '@ephox/katamari';
+
 import * as TabbingTypes from './TabbingTypes';
-import { FieldSchema, FieldProcessorAdt } from '@ephox/boulder';
-import { Fun, Option } from '@ephox/katamari';
 
 export default TabbingTypes.create(
   FieldSchema.state('cyclic', Fun.constant(false))