+++ src/main/ts/ephox/alloy/keying/EscapingType.ts
@@ -3,15 +3,12 @@ import { Fun, Option } from '@ephox/katamari';
 import * as Keys from '../alien/Keys';
 import { NoState, Stateless } from '../behaviour/common/BehaviourState';
 import * as Fields from '../data/Fields';
+import { EscapingConfig, KeyRuleHandler } from '../keying/KeyingModeTypes';
 import * as KeyMatch from '../navigation/KeyMatch';
 import * as KeyRules from '../navigation/KeyRules';
 import * as KeyingType from './KeyingType';
-import { KeyRuleHandler, EscapingConfig } from '../keying/KeyingModeTypes';
 
 // NB: Tsc requires AlloyEventHandler and AlloyComponent to be imported here.
-import { AlloyEventHandler } from '../api/events/AlloyEvents';
-import { AlloyComponent } from '../api/component/ComponentApi';
-
 const schema = [
   Fields.onStrictKeyboardHandler('onEscape')
 ];
@@ -20,11 +17,10 @@ const doEscape: KeyRuleHandler<EscapingConfig, Stateless> = (component, simulate
   return escapeConfig.onEscape()(component, simulatedEvent);
 };
 
-const getRules: () => Array<KeyRules.KeyRule<EscapingConfig, Stateless>> = Fun.constant([
+const getKeydownRules: () => Array<KeyRules.KeyRule<EscapingConfig, Stateless>> = Fun.constant([
   KeyRules.rule(KeyMatch.inSet(Keys.ESCAPE()), doEscape)
 ]);
 
-const getEvents = Fun.constant({ });
-const getApis = Fun.constant({ });
+const getKeyupRules = Fun.constant([ ]);
 
-export default KeyingType.typical(schema, NoState.init, getRules, getEvents, getApis, Option.none());
+export default KeyingType.typical(schema, NoState.init, getKeydownRules, getKeyupRules, Option.none());