+++ src/main/ts/ephox/alloy/keying/ExecutionType.ts
@@ -1,19 +1,16 @@
-import { FieldSchema, FieldProcessorAdt } from '@ephox/boulder';
-import { Fun, Option } from '@ephox/katamari';
+import { FieldSchema } from '@ephox/boulder';
+import { Option } from '@ephox/katamari';
 
 import * as EditableFields from '../alien/EditableFields';
 import * as Keys from '../alien/Keys';
+import { AlloyComponent } from '../api/component/ComponentApi';
 import { NoState, Stateless } from '../behaviour/common/BehaviourState';
+import { NativeSimulatedEvent } from '../events/SimulatedEvent';
+import { ExecutingConfig, KeyRuleHandler } from '../keying/KeyingModeTypes';
 import * as KeyMatch from '../navigation/KeyMatch';
 import * as KeyRules from '../navigation/KeyRules';
 import * as KeyingType from './KeyingType';
 import * as KeyingTypes from './KeyingTypes';
-import { ExecutingConfig, KeyRuleHandler } from '../keying/KeyingModeTypes';
-
-import { AlloyComponent } from '../api/component/ComponentApi';
-import { SugarEvent } from '../alien/TypeDefinitions';
-import { EventFormat, SimulatedEvent, NativeSimulatedEvent } from '../events/SimulatedEvent';
-import { AlloyEventHandler } from '../api/events/AlloyEvents';
 
 const schema = [
   FieldSchema.defaulted('execute', KeyingTypes.defaultExecute),
@@ -27,7 +24,7 @@ const execute: KeyRuleHandler<ExecutingConfig, Stateless> = (component: AlloyCom
   return executeConfig.execute()(component, simulatedEvent, component.element());
 };
 
-const getRules = (component: AlloyComponent, simulatedEvent: NativeSimulatedEvent, executeConfig: ExecutingConfig, executeState: Stateless): Array<KeyRules.KeyRule<ExecutingConfig, Stateless>> => {
+const getKeydownRules = (component: AlloyComponent, simulatedEvent: NativeSimulatedEvent, executeConfig: ExecutingConfig, executeState: Stateless): Array<KeyRules.KeyRule<ExecutingConfig, Stateless>> => {
   const spaceExec = executeConfig.useSpace() && !EditableFields.inside(component.element()) ? Keys.SPACE() : [ ];
   const enterExec = executeConfig.useEnter() ? Keys.ENTER() : [ ];
   const downExec = executeConfig.useDown() ? Keys.DOWN() : [ ];
@@ -40,7 +37,10 @@ const getRules = (component: AlloyComponent, simulatedEvent: NativeSimulatedEven
   ] : [ ]);
 };
 
-const getEvents = Fun.constant({ });
-const getApis = Fun.constant({ });
+const getKeyupRules = (component: AlloyComponent, simulatedEvent: NativeSimulatedEvent, executeConfig: ExecutingConfig, executeState: Stateless): Array<KeyRules.KeyRule<ExecutingConfig, Stateless>> => {
+  return executeConfig.useSpace() && !EditableFields.inside(component.element()) ? [
+    KeyRules.rule(KeyMatch.inSet(Keys.SPACE()), KeyingTypes.stopEventForFirefox)
+  ] : [ ]
+};
 
-export default KeyingType.typical(schema, NoState.init, getRules, getEvents, getApis, Option.none());
+export default KeyingType.typical(schema, NoState.init, getKeydownRules, getKeyupRules, Option.none());