+++ src/main/ts/ephox/alloy/keying/FlowType.ts
@@ -69,7 +69,7 @@ const doEscape: KeyRuleHandler<FlowConfig, Stateless>  = (component, simulatedEv
   return flowConfig.onEscape()(component, simulatedEvent);
 };
 
-const getRules = (_component, _se, flowConfig, _flowState): Array<KeyRules.KeyRule<FlowConfig, Stateless>> => {
+const getKeydownRules = (_component, _se, flowConfig, _flowState): Array<KeyRules.KeyRule<FlowConfig, Stateless>> => {
   const westMovers = Keys.LEFT().concat(flowConfig.allowVertical() ? Keys.UP() : [ ]);
   const eastMovers = Keys.RIGHT().concat(flowConfig.allowVertical() ? Keys.DOWN() : [ ]);
   return [
@@ -81,8 +81,8 @@ const getRules = (_component, _se, flowConfig, _flowState): Array<KeyRules.KeyRu
   ];
 };
 
-const getEvents = Fun.constant({ });
+const getKeyupRules: () => Array<KeyRules.KeyRule<FlowConfig, Stateless>> = Fun.constant([
+  KeyRules.rule(KeyMatch.inSet(Keys.SPACE()), KeyingTypes.stopEventForFirefox)
+])
 
-const getApis = Fun.constant({ });
-
-export default KeyingType.typical(schema, NoState.init, getRules, getEvents, getApis, Option.some(focusIn));
+export default KeyingType.typical(schema, NoState.init, getKeydownRules, getKeyupRules, Option.some(focusIn));